backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pytesseract==0.3.13
python-dateutil==2.8.2
supabase==2.3.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
